% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_from_convert.R
\name{calc_from_convert}
\alias{calc_from_convert}
\title{Convert calculation list to a specific format}
\usage{
calc_from_convert(x)
}
\arguments{
\item{x}{A list of calculations.}
}
\value{
A formatted list of calculations.
}
\description{
Convert calculation list to a specific format
}
