% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{calculation}
\alias{calculation}
\title{Calculation Class}
\description{
Represents a calculation with associated parameters, filters, and sub-calculations.
}
\section{Methods}{

\describe{
\item{\code{add_sub_calculation(sub_calculation, name)}}{Add a sub-calculation.}
\item{\code{data_clone(...)}}{Clone the data.}
}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{function_name}}{Character. The name of the function.}

\item{\code{parameters}}{List. The parameters for the calculation.}

\item{\code{calculated_from}}{Character vector. Sources from which the calculation is derived.}

\item{\code{is_recalculable}}{Logical. Indicates if the calculation is recalculable.}

\item{\code{sub_calculations}}{List. A list of sub-calculations.}

\item{\code{type}}{Character. The type of calculation.}

\item{\code{filter_conditions}}{List. The conditions used to filter data.}

\item{\code{filters}}{List. The filters applied to the data.}

\item{\code{name}}{Character. The name of the calculation instance.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-calculation-new}{\code{calculation$new()}}
\item \href{#method-calculation-add_sub_calculation}{\code{calculation$add_sub_calculation()}}
\item \href{#method-calculation-data_clone}{\code{calculation$data_clone()}}
\item \href{#method-calculation-clone}{\code{calculation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculation-new"></a>}}
\if{latex}{\out{\hypertarget{method-calculation-new}{}}}
\subsection{Method \code{new()}}{
Initialize the calculation class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calculation$new(
  function_name = "",
  parameters = list(),
  calculated_from = c(),
  is_recalculable = TRUE,
  sub_calculations = list(),
  type = "",
  filter_conditions = list(),
  filters = list(),
  name = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{function_name}}{Character. The name of the function. Default is an empty string.}

\item{\code{parameters}}{List. The parameters for the calculation. Default is an empty list.}

\item{\code{calculated_from}}{Character vector. Sources from which the calculation is derived. Default is an empty vector.}

\item{\code{is_recalculable}}{Logical. Indicates if the calculation is recalculable. Default is \code{TRUE}.}

\item{\code{sub_calculations}}{List. A list of sub-calculations. Default is an empty list.}

\item{\code{type}}{Character. The type of calculation. Default is an empty string.}

\item{\code{filter_conditions}}{List. The conditions used to filter data. Default is an empty list.}

\item{\code{filters}}{List. The filters applied to the data. Default is an empty list.}

\item{\code{name}}{Character. The name of the calculation instance. Default is an empty string.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculation-add_sub_calculation"></a>}}
\if{latex}{\out{\hypertarget{method-calculation-add_sub_calculation}{}}}
\subsection{Method \code{add_sub_calculation()}}{
Add a sub-calculation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calculation$add_sub_calculation(sub_calculation, name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sub_calculation}}{An object representing the sub-calculation to add.}

\item{\code{name}}{Character. The name of the sub-calculation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculation-data_clone"></a>}}
\if{latex}{\out{\hypertarget{method-calculation-data_clone}{}}}
\subsection{Method \code{data_clone()}}{
Clone the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calculation$data_clone(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional methods to add to the function.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of the calculation class with the same data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-calculation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calculation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
