% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculationSystemR-package.R
\docType{package}
\name{calculationSystemR-package}
\alias{calculationSystemR}
\alias{calculationSystemR-package}
\title{calculationSystemR: What the Package Does (One Line, Title Case)}
\description{
Calculation system package that contains an R6 class.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
