% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_filter.R
\name{check_filter}
\alias{check_filter}
\title{Check and update filter object parameters}
\usage{
check_filter(filter_obj)
}
\arguments{
\item{filter_obj}{A filter object to check and update.}
}
\value{
The updated filter object.
}
\description{
Check and update filter object parameters
}
