% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_df_from_calc_from.R
\name{find_df_from_calc_from}
\alias{find_df_from_calc_from}
\title{Find data frame from calculation list}
\usage{
find_df_from_calc_from(x, column)
}
\arguments{
\item{x}{A list of calculations.}

\item{column}{The column name to search for.}
}
\value{
The name of the data frame associated with the column.
}
\description{
Find data frame from calculation list
}
