% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instat_calculation.R
\name{instat_calculation}
\alias{instat_calculation}
\title{instat_calculation Class}
\description{
instat_calculation Class

instat_calculation Class
}
\details{
A class to store calculations.
}
\section{Methods}{

\describe{
\item{\code{get_dependencies(depens = c())}}{Get Dependencies.}
\item{\code{data_clone(...)}}{Clone the data.}
}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{function_exp}}{A string passed directly to one of dplyr functions.}

\item{\code{type}}{The type of calculation.}

\item{\code{name}}{The name of the calculation instance.}

\item{\code{result_name}}{The name for the output produced by the calculation.}

\item{\code{result_data_frame}}{The data frame that the output should go to.}

\item{\code{manipulations}}{A list of calculations to be performed before sub_calculations and the main calculation.}

\item{\code{sub_calculations}}{A list of calculations to be performed after manipulations.}

\item{\code{calculated_from}}{A list of columns the calculation depends on.}

\item{\code{save}}{An integer indicating whether the calculation and result should be saved.}

\item{\code{before}}{A boolean indicating if the calculation should be performed before others.}

\item{\code{adjacent_column}}{The name of the adjacent column.}

\item{\code{param_list}}{Additional parameters to read into the calculation, e.g., \code{.drop = TRUE} for \code{by} calculation.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-instat_calculation-new}{\code{instat_calculation$new()}}
\item \href{#method-instat_calculation-data_clone}{\code{instat_calculation$data_clone()}}
\item \href{#method-instat_calculation-get_dependencies}{\code{instat_calculation$get_dependencies()}}
\item \href{#method-instat_calculation-clone}{\code{instat_calculation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-instat_calculation-new"></a>}}
\if{latex}{\out{\hypertarget{method-instat_calculation-new}{}}}
\subsection{Method \code{new()}}{
Initialise the instat_calculation class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{instat_calculation$new(
  function_exp = "",
  type = "",
  name = "",
  result_name = "",
  result_data_frame = "",
  manipulations = list(),
  sub_calculations = list(),
  calculated_from = list(),
  save = 0,
  before = FALSE,
  adjacent_column = "",
  param_list = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{function_exp}}{A string passed directly to one of dplyr functions.}

\item{\code{type}}{The type of calculation.}

\item{\code{name}}{The name of the calculation instance.}

\item{\code{result_name}}{The name for the output produced by the calculation.}

\item{\code{result_data_frame}}{The data frame that the output should go to.}

\item{\code{manipulations}}{A list of calculations to be performed before sub_calculations and the main calculation.}

\item{\code{sub_calculations}}{A list of calculations to be performed after manipulations.}

\item{\code{calculated_from}}{A list of columns the calculation depends on.}

\item{\code{save}}{An integer indicating whether the calculation and result should be saved.}

\item{\code{before}}{A boolean indicating if the calculation should be performed before others.}

\item{\code{adjacent_column}}{The name of the adjacent column.}

\item{\code{param_list}}{Additional parameters to read into the calculation, e.g., \code{.drop = TRUE} for \code{by} calculation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-instat_calculation-data_clone"></a>}}
\if{latex}{\out{\hypertarget{method-instat_calculation-data_clone}{}}}
\subsection{Method \code{data_clone()}}{
Clone the data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{instat_calculation$data_clone(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional methods to add to the function.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of the instat_calculation class with the same data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-instat_calculation-get_dependencies"></a>}}
\if{latex}{\out{\hypertarget{method-instat_calculation-get_dependencies}{}}}
\subsection{Method \code{get_dependencies()}}{
Get dependencies
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{instat_calculation$get_dependencies(depens = c())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{depens}}{A vector of dependencies.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of dependencies.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-instat_calculation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-instat_calculation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{instat_calculation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
