% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculationSystemR-package.R
\docType{package}
\name{instatCalculations-package}
\alias{instatCalculations}
\alias{instatCalculations-package}
\title{instatCalculations: Calculation system used in R-Instat}
\description{
Functions used in the calculation system used in R-Instat
}
\author{
\strong{Maintainer}: Lily Clements \email{lily@idems.international}

Authors:
\itemize{
  \item Danny Parsons
}

}
\keyword{internal}
