% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{calculation}
\alias{calculation}
\title{calculation Class}
\description{
An example R6 class for performing calculations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{function_name}}{The name of the function.}

\item{\code{parameters}}{The parameters for the calculation.}

\item{\code{calculated_from}}{A vector of sources from which the calculation is derived.}

\item{\code{is_recalculable}}{A boolean indicating if the calculation is recalculable.}

\item{\code{sub_calculations}}{A list of sub-calculations.}

\item{\code{type}}{The type of calculation.}

\item{\code{filter_conditions}}{The conditions used to filter data.}

\item{\code{filters}}{The filters applied to the data.}

\item{\code{name}}{The name of the calculation instance.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-calculation-new}{\code{calculation$new()}}
\item \href{#method-calculation-clone}{\code{calculation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculation-new"></a>}}
\if{latex}{\out{\hypertarget{method-calculation-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calculation$new(
  function_name = "",
  parameters = list(),
  calculated_from = c(),
  is_recalculable = TRUE,
  sub_calculations = list(),
  type = "",
  filter_conditions = list(),
  filters = list(),
  name = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{function_name}}{The name of the function. Default is an empty string.}

\item{\code{parameters}}{A list of parameters for the calculation. Default is an empty list.}

\item{\code{calculated_from}}{A vector of sources from which the calculation is derived. Default is an empty vector.}

\item{\code{is_recalculable}}{A boolean indicating if the calculation is recalculable. Default is TRUE.}

\item{\code{sub_calculations}}{A list of sub-calculations. Default is an empty list.}

\item{\code{type}}{The type of calculation. Default is an empty string.}

\item{\code{filter_conditions}}{The conditions used to filter data. Default is an empty list.}

\item{\code{filters}}{The filters applied to the data. Default is an empty list.}

\item{\code{name}}{The name of the calculation instance. Default is an empty string.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-calculation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calculation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
